requirements.txt
chromadb==0.4.23
ctransformers==0.2.27
InstructorEmbedding==1.0.1
langchain==0.1.9
langchain-community==0.0.22
llama-cpp-python==0.2.20
pypdfium2==4.27.0
pyyaml==6.0.1
sentence-transformers==2.3.1
streamlit==1.31.1
transformers==4.38.1
torch==2.2.0
types-PyYAML
ctransformers[cuda]
# separately install cu118:  pip install torch==2.2.0 torchvision==0.17.0 torchaudio==2.2.0 --index-url https://download.pytorch.org/whl/cu118